Fix bug in eq_stmtkind / eq_stmt_with_location.
The previous implementation directly compared the sids of the nodes
a GOTO statement would jump to.
This is a problem, however:
The sids can change, if, e.g. another function is added to the source code.
Thus, a function with a goto/loop/if would be considered as changed,
even if this is not the case.