Create and use a method Goblintutil.create_var.
This method is intended to make the id of variables created
by Goblint independend of how many varibales have been created before.
This is feature is needed for incremental analyses, in order to allow
us to reuse results, as semantically same variables will have the same ids.
Thus, the contexts and values in which they appear can be reused.