Use custom hashmap module in SolverStats instead of Hashtbl.
Use a custom hashmap module, with custom compare and hash functions in
SolverStats. This fixes an out-of-memory error (stack overflow) that can
occur when using the built-in functions of Hashtbl that rely on
Pervasives.compare.

In particular, this out-of-memory error was more likely to appear with
incremental analysis, as some identical objects may appear
at different memory addresses. Pervasives.compare in this case may
not be able to detect that two cyclical structures are identical.