restore SetDomain.HoarePO.hash to be based on just the elements again, fixes hashcons
In this branch this was changed to polymorphic hash which gives different values if the set representation changes.
This in turn leads to values landing in different buckets in the weak hashtable if hashcons is enabled and
therefore HConsed.equal will often fail for values that are actually Base.equal.

Running this on figlet with hashcons on took only 4m 51s (S.Dom.equal 1.076 s) instead of 7m 14s (S.Dom.equal 40.744 s) with hashcons off.
Running with the code from the previous debug commit took 11m 43s with hashcons on.