initialize locals with bot, warn on access (undefined behavior)
Accessing uninitialized objects is undefined behavior and we should warn about it!
The access might happen in [get] or in [do_offs] in some sub-expression in the above [eval_rv].
Since bot propagates and we don't want to warn for all sub-expressions we only warn here.
Previously we only gave sound results for programs without undefined behavior, so yielding bot for accessing an uninitialized array was considered ok.
Now we want to warn about about any access to uninitialized memory and treat it as top/Unknown instead of bot/Deadcode.
We could just change bot to top here, but then we would not see which values are transitively uninitialzed (int x; int y = x; // y should also be bot).
So we only warn here, keep bot, and instead treat bot as top when branching. Locals are now initialized with bot instead of top.
This way bot for variables means uninitalized (before we only had bot inside arrays and blobs).
To be sound we adjust:
- [branch] such that it falls through and meets [invariant] in both branches instead of raising Deadcode if eval_rv cond is bot
- asserts to return UNKNOWN
-  calls of expressions that evaluate to bot TODO can this happen?