Fix {track,trace}_{on,off}.sh to work with dune
Was complaining about some files being present, and others not (see blow). Run `make gen` to solve this issue.

Error: Some of the targets of this fallback rule are present in the source
tree, and some are not. This is not allowed. Either none of the targets must
be present in the source tree, either they must all be.

The following targets are present:
- src/config.ml
- src/version.ml

The following targets are not:
- src/goblint.ml
Makefile:4: recipe for target 'native' failed
make: *** [native] Error 1