Base.invariant: Use tuple of type (ID.t * DefExc.t) for computation.
Intervals cannot represent the abstract value for true accurately.
Thus, extending the abstract value of c: ID.t in the invariant code
with DefExc.t to c: ID.t * DefExc.t helps with precision in true branches.

This patch is extremly ugly and should be refactored.
E.g. by creating a module for ID.t * DefExc.t and using it in the
invariant code.