Make IntDomTuple configurable, possibly independent of Goblint-config.
Move code of IntDomTuple to functor IntDomTupleFunc, which
takes a configuration of which domains should be active.
Create a module IntDomeTuple that is using the Goblint configuration.

Now, we might reuse IntDomTuple in code parts
(e.g. in the future: branching/invariant code) where we possibly want to
use a different configuration.