Invariant: Use DefExc in invariant code independent of configuration.
The interval domain alone cannot accurately represent the set of concrete
values that evaluate to true (i.e. the set of all integers except 0).
The invariant code previsouly could not restrict the abstract values
for the interval domain when used by itself in "true"-branches
(unless using an unsound abstract value for "true").

This patch ensures that in the invariant code the DefExc domain is always used,
irregardless of the used Goblint configuration. Thus, the uses of
"to_bool" in the invariant code should be able to recover the truth values
in more cases. This should yield more accuracy in these cases.