Def_exc: In apply_range handle Int64 overflow
When performing an injective operation on `Definite and Excluded  apply_range is used to determine the new range.
To do this, the definite value is applied to the minimum and the maximum of the range. Unfortunately, these computations on Int64 can overflow for alues of type long.

This leads e.g. to `Excluded{}[-63, -63] - 2 == Excluded {} [0,64] which is unsound. To prevent this, if the range is such that the computation may overflow, we set the range to top_range instead.
References #93