Enums: Do not perform binary arithmetic operations on sets with multiple elements.
In some special cases, like multiplication with a definite value of
one or zero, the precise result is still provided.

Removing this logic for the general case of binary operations with sets of more than one value
is aimed at improving analysis performance and making the domain implementation more simple.