Library Analyis Combine: Perform update for reachable memory blocks separately for each argument.
Perform update for reachable memory blocks separately for each argument.
This is necessary because strange behavior of the address domain were the
the result of separate calls to reachable_vars is not subsumed by a call with
the joined value. In particular this occurs when a call to function contains
a pointer to a struct, as well as a pointer to the first field in the struct.

Additionally, the mallocWrapperTypeBasedAnalysis now assocites the correct type to
allocated memory blocks.