Handle writes to globals in the library analysis.
get_symbolic_address will now return the global if the concrete value was a global. Globals and args are now passed both to update_lvals at the same time. This is necessary so that updated pointers in the local state really can be made to point to all reachable locations