RichVarinfo: Separate BiDirectional mapping into own module, as only the MallocWrapperAnalysis currently needs it.
Only the BiDirectional mappings need and can profit from the describe_varinfo vi t funtion,
so only these BiDirectional mappings requires such a module as an input.
BiDirectional mappings reqister themselves into a Collection; this Collection is quried by functions
that output varinfos that might be contained in such a mapping to enrich the output with a description.