Destabilize the entry points of force-reanalyzed functions, even when incremental.reluctant.on is activated.
Force-reanalyze is intended to be used when functions are to be reanalyzed with different precision, e.g. differnt int-domains activated.
Reluctant analysis (as-is) does not work for this kind of re-analysis, as the entry transfer function is not re-evaluated in the reluctant mode. Thus, this commit disables reluctant analysis for functions that are contained in the list of force analyzed functions.