Add unkowns from partially changed functions to obsolete_prim when reluctant is off.
The set `obsolete_prim` is eagerly destablized; it does not need to be collected when reluctant is on.
When reluctant is on and at the same time, some function is in force-reanalyze,
this function will be considered a completely changed function, and not as partially changed.
Therefore the unknowns belonging to its function entry node are collected in
`obsolete_entry` beforehand; which will be eagerly destablized as required.