Do not store and load version_map.
This map was intended to keep the latest version of globals (functions and variables) that actually appeared in the computed analysis restuts.
Storing this map is not (or no longer) necessary, as the Cil.file that is stored already contains all relevant renamings. Only the maximum ids still have to be stored.