Add node as optional parameter passed to ask for transformations.
When a node is passed, the ask will use this as the current node.
This is used by the EvalAssert transformation. Passing the node there
allows the var_is_in_scope function to return the right result. In turn,
assertions on local variables are now created by the evalAssert transformation.