Allow queries from modular anlaysis to non-modular analysis during combine_env.
Improves the handling when a function that is to be analyzed modularly is called from a function that is analyzed non-modularly.
The ModularWrapper is changed such that it keeps `Left entries for analyses that should be able to be queried, but for which the transfer function should actually not be computed.
Then, for the analysis of the callee, the non-modular analysis are completely reomved, while for the combine_env, the non-modular anlaysis are kept as queriable.