Make join/smart_join/widen/smart_widen preserve equal inputs.
Issue #1091 was caused by the join/widen (and corresponding smart operations)
for unions in ValueDomain not return the same element.
The optimization in PMap that returns the unchanged map when two
physically identical maps were joined was thus yielding a different value in the
from-scratch run compared to the incremental run.