Add ValueDomain.try_meet, meet that may fail for unions with different active fields, use it in BaseInvariant.
The refinement by BaseInvariant to handle branching used VD.meet, which
caused issue #1105, due to the behavior of UnionDomain.meet:
The meet of unions with different active fields can result in a bottom value.
This led to e.g. spuriously dead code.

This commit adds a function try_meet to the valueDomain, the union domain,
and those domains that may contain a union domain element: structs, arrays, blobs.
The function try_meet in the union domain throws an exception in the
case the fields of the unions for which a meet was tried did not coincide.
In case this exception occurs, the BaseInvariant does not perform
refinement.