changed leq in linearTwoVarEqualityDomain.apron.ml
changed the condition for implication of a single equality in 
the less or equal function.
A conjunction of equalities might represent a set of states that is a subset of another set of states and its corresponding representation as a conjunction of equalities.
In particular, an assignment to a variable only consisting of a constant offset implies, say, the trivial equality, which is satisfied for any assignment to the same variable. 
The change correctly classifies this kind of implication.