C2PO: Adapt C2PO such that is does not read attributes from richvarinfo vars.
The C2PO analysis maps some varinfos to varinfos via the RichVarinfo module. The code used to read some attributes from these variables from the result varinfos.
Thus, it required a change where the all the attributes of resulting rich varinfos can be specified. With this change, this is no longer necessary.