Parallelism support: PoC
In order to maintain backwards compatibility, we need to provide
stubs for functionality needed for parallelism when no libraries
provide this. Also, to keep possibly diverging parts to a minimum,
the utilities for parallelism are kept in a separate module.

This commit is a sketch of how we could go about it. In particular,
it is not clear if `ppx_optcomp` is the best solution for this.