Use-after-free analysis: Improve checking of accesses to lvals.
Given, e.g., a statement

lval = rval;

There may be two types of use-after-free errors that can occur for the lval:

1. The memory location to which lval evaluates has been freed already.
2. The memory location(s) inside rval that are dereferenced to evaluate rval have been freed already.

Previously, the check implemented for lvals did not consistently check for both types of errors.
In case a dereference was at the top-level of the lval, only the second type of error was checked, but not recursively for sub-lvals.

This is changed with this PR.
The 1. check is now performed, whether the lhost is a variable or a memory dereference.
The second check is performed using a recursive call of the mutually recursive function `warn_exp_might_contain_freed`.