Implement ana.apron.invariant.diff-box directly in domain
This should be faster, because it doesn't require box domain analysis in parallel.
The domain is converted to box and back only during invariant generation.

It's also more precise in some sense because it can exclude box-like constraints, which cannot be inferred by box alone.
But maybe that's also a downside because it can exclude constraints, which still need the more precise domain.